fix(pulam-web): nix run .#pulam-web build — split port helpers into surface-nix-host-free ports.ts
The Vite client build (hence `nix run .#pulam-web`) failed at construction:
vite.config.ts imports the port defaults from config.ts, which imports
@kolu/surface-nix-host, whose barrel uses extensionless relative imports
(`export … from "./arch"`). The workspace bundler-mode tsc/vitest resolve
those, but Node's NATIVE ESM (which loads vite.config) does not →
ERR_MODULE_NOT_FOUND, so the client bundle never built.

Extract DEFAULT_PORT/DEFAULT_CLIENT_PORT/parsePort into ports.ts (stdlib-only,
no surface-nix-host import); config.ts re-exports them so server consumers are
unchanged; vite.config.ts imports from ports.ts. `vite build` now succeeds.

Caught by running nix run .#pulam-web live on a pu box (typecheck/vitest use
bundler resolution, so they never surfaced it).
