+++ packages/pulam-web/src/server/config.ts
@@ -46,47 +46,13 @@ export const PULAM_AGENT_DRVS_ENV = "PULAM_AGENT_DRVS_JSON";
  *  host with one kaval needs no entry (discovery is unambiguous). */
 export const PULAM_WEB_KAVAL_SOCKETS_ENV = "PULAM_WEB_KAVAL_SOCKETS";
 
-/** The HTTP+WebSocket port. `4800` is pulam-web's default (the `48` echoes the
- *  R4.8 epic). Override with `PULAM_WEB_PORT`. */
-export const DEFAULT_PORT = 4800;
-
-/** The dev client (Vite) default port. `5800` pairs with `DEFAULT_PORT`'s `48`.
- *  Override with `PULAM_WEB_CLIENT_PORT`. */
-export const DEFAULT_CLIENT_PORT = 5800;
-
-/**
- * Parse a port env var, using `fallback` ONLY when the var is unset/empty.
- *
- * Fail-fast, no silent fallback: a present-but-malformed value (`abc`, `12.5`,
- * `99999`, or an explicit `0` — which would bind an arbitrary OS-assigned port,
- * never what a config author means) THROWS, naming the var, rather than quietly
- * collapsing to the default the way `Number(env) || fallback` does (the F6 bug:
- * `Number("abc") || 4800` and `Number("0") || 4800` both silently yield 4800).
- * A valid integer in `1..65535` is returned as-is.
- */
-export function parsePort(
-  varName: string,
-  raw: string | undefined,
-  fallback: number,
-): number {
-  if (raw === undefined || raw.trim() === "") return fallback;
-  const trimmed = raw.trim();
-  // Reject anything that isn't a plain non-negative integer literal up front —
-  // `Number` would accept "12.5", "0x10", "1e3", " 80 " and leading/trailing
-  // junk via coercion, none of which is a port a config author typed on purpose.
-  if (!/^\d+$/.test(trimmed)) {
-    throw new Error(
-      `${varName}: invalid port ${JSON.stringify(raw)} — must be an integer in 1..65535.`,
-    );
-  }
-  const port = Number(trimmed);
-  if (port < 1 || port > 65535) {
-    throw new Error(
-      `${varName}: port ${port} out of range — must be an integer in 1..65535 (0 is rejected: it would bind an arbitrary OS-assigned port).`,
-    );
-  }
-  return port;
-}
+// The port constants + the strict `parsePort` live in `ports.ts` — a module that
+// imports NOTHING from `@kolu/surface-nix-host`, so `vite.config.ts` can read the
+// dev-proxy port defaults without dragging this module's drv-resolver (and its
+// extensionless-barrel imports that native ESM can't load) into the Vite config
+// load. Re-exported here so server-side code still reads everything from
+// `config.ts`. See `ports.ts` for the full rationale.
+export { DEFAULT_CLIENT_PORT, DEFAULT_PORT, parsePort } from "./ports.ts";
 
 /** Read + split the static host set from `PULAM_WEB_HOSTS`. Each entry is
  *  trimmed; empty entries (a trailing comma, a double comma) are dropped. An
